Refactor code for type safety and performance improvements
- Changed variable types from int to size_t where appropriate in Byzantine detector and LLM model storage to enhance type safety.
- Updated gradient selection logic in Krum detector to ensure correct type casting.
- Modified data loader to explicitly cast batch size to int for clarity.
- Adjusted Llama tokenizer to use static_cast for length and buffer size conversions.
- Enhanced Adam and AdamW optimizers by ensuring bias correction terms are explicitly cast to float.
- Removed unused function getModelPathProviderFn in LoRATrainingService.
- Optimized hyperparameter initialization in LoRATrainingService for clarity and consistency.
- Improved metrics calculations in LoRATrainingService by using static_cast for total steps and current steps.
- Refined LoRASecurityValidator to ensure kurtosis calculations are type-safe.
- Updated MultiLoRAManager to correctly cast alpha from string to size_t.
- Enhanced QueryEngine by replacing push_back with emplace_back for better performance and clarity in various locations.
- Improved RocksDBWrapper by using emplace_back for adding listeners and chunk keys, and reserving space for vectors to optimize performance.

+++ src/llm/byzantine_detector.cpp
@@ -156,7 +156,7 @@ GradientStatistics MedianDetector::computeStatistics(
         
         // Compute mean
         float sum = 0.0f;
-        int count = 0;
+        size_t count = 0;
         for (const auto& tensor : gradients) {
             sum += std::accumulate(tensor.data.begin(), tensor.data.end(), 0.0f);
             count += tensor.data.size();
@@ -298,7 +298,7 @@ std::vector<std::string> KrumDetector::selectKrumGradients(
     }
     
     // For each gradient, compute sum of distances to k closest gradients
-    int k = n - max_byzantine_shards_ - 2;  // Number of closest neighbors to consider
+    int k = static_cast<int>(n) - max_byzantine_shards_ - 2;  // Number of closest neighbors to consider
     k = std::max(1, k);
     
     std::vector<std::pair<float, std::string>> scores;
@@ -357,7 +357,7 @@ DetectionResult KrumDetector::detectByzantineShards(
     }
     
     // Number of gradients to select
-    int m = n - max_byzantine_shards_ - 2;
+    int m = static_cast<int>(n) - max_byzantine_shards_ - 2;
     
     try {
         auto selected = selectKrumGradients(shard_gradients, m);
